WAF增加多个动作

+++ pkg/serverconfigs/firewallconfigs/http_firewall_action_get_302.go
@@ -3,5 +3,5 @@
 package firewallconfigs
 
 type HTTPFirewallGet302Action struct {
-	Life int64 `yaml:"life" json:"life"`
+	Life int32 `yaml:"life" json:"life"`
 }
